+++ SecureAccessBLE/Classes/Manager/BLEComManager.swift
@@ -20,6 +20,23 @@ public protocol BLEManagerDelegate: class {
     func bleDidReceivedServiceTriggerForStatus(_ status: ServiceGrantTriggerStatus?, error: String?)
     
     /**
+     BLE reports if a connection attempt succeeded
+     
+     - parameter communicator: The communicator object
+     - parameter sid: The SID the connection is made to
+     */
+    func bleDidConnectSid(_ manager: BLEComManager, sid: SID)
+    
+    /**
+     BLE reports if a connection attempt failed
+     
+     - parameter manager: The manager object
+     - parameter sid: The SID the connection should have made to
+     - parameter error: Describes the error
+     */
+    func bleDidFailToConnectSid(_ manager: BLEComManager, sid: SID, error: Error?)
+    
+    /**
      BLE changed connection status
      
      - parameter isConnected: currently connected or not
@@ -39,6 +56,11 @@ public protocol BLEManagerDelegate: class {
      - parameter oldSids: the lost old sids as array
      */
     func bleDidLostSidIds(_ oldSids: [SID])
+    
+    /**
+     BLE reports blob needs to be updated, because the user is sending an out of date blob token
+     */
+    func blobIsOutdated()
 }
 
 // MARK: - Extension point for BLEmanager delegate
@@ -50,6 +72,23 @@ public extension BLEManagerDelegate {
      - parameter error:  error description
      */
     func bleDidReceivedServiceTriggerForStatus(_ status: ServiceGrantTriggerStatus?, error: String?) {}
+    
+    /**
+     BLE reports if a connection attempt succeeded
+     
+     - parameter communicator: The communicator object
+     - parameter sid: The SID the connection is made to
+     */
+    func bleDidConnectSid(_ manager: BLEComManager, sid: SID) {}
+    
+    /**
+     BLE reports if a connection attempt failed
+     
+     - parameter manager: The manager object
+     - parameter sid: The SID the connection should have made to
+     - parameter error: Describes the error
+     */
+    func bleDidFailToConnectSid(_ manager: BLEComManager, sid: SID, error: Error?) {}
     
     /**
      BLE changed connection status
@@ -71,6 +110,7 @@ public extension BLEManagerDelegate {
      - parameter oldSids: the lost old sids as array
      */
     func bleDidLostSidIds(_ oldSids: [SID]) {}
+
 }
 
 /**
@@ -163,16 +203,15 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
     ///The connection state
     open var isConnected: Bool {
         let connectedState = self.currentConnectionState == .connected
-        //let encryptionNoState = self.currentEncryptionState == .NoEncryption
         let encryptionEstablished = self.currentEncryptionState == .encryptionEstablished
-        //debugPrint("connected? \(self.currentConnectionState) NoEncryption?: \(encryptionNoState) EncryptionEstablished?: \(encryptionEstablished)")
-        return connectedState && encryptionEstablished//(encryptionNoState || encryptionEstablished)
+        print("connected? \(self.currentConnectionState) EncryptionEstablished?: \(encryptionEstablished)")
+        return connectedState && encryptionEstablished
     }
     
     /// Connection state, default as .Notconnected
     fileprivate var currentConnectionState = ConnectionState.notConnected {
         didSet {
-            debugPrint("State changed now: \(currentConnectionState)")
+            print("State changed now: \(currentConnectionState)")
         }
     }
     
@@ -181,9 +220,9 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
     /// Chanllenger object
     fileprivate var challenger: BLEChallengeService?
     ///  The communicator objec
-    fileprivate var communicator: SIDCommunicator?
+    private let communicator: SIDCommunicator
     /// DeviceId as String came from Userspace.Booking
-    open var deviceId: String = ""
+    open var leaseId: String = ""
     
     /// LeaseToken Id as String came from SecureAccess.blob
     open var leaseTokenId: String = ""
@@ -237,11 +276,11 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
             self.currentEncryptionState = .shouldEncrypt
         }
         self.transporter = BLEScanner()
-        super.init()
         self.communicator = SIDCommunicator.init()
-        self.communicator?.transporter = self.transporter
-        self.communicator?.resetFoundSids()
-        self.communicator?.delegate = self
+        self.communicator.transporter = self.transporter
+        self.communicator.resetFoundSids()
+        super.init()
+        self.communicator.delegate = self
     }
     
     /**
@@ -264,7 +303,7 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
      Deinit point
      */
     deinit {
-        debugPrint("Will be both BLE and Comm. disconnected!")
+        print("Will be both BLE and Comm. disconnected!")
         self.disconnect()
         self.disconnectTransporter()
     }
@@ -275,8 +314,8 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
      - returns: Transfer busy
      */
     open func transferIsBusy() -> Bool {
-        let transferIsBusy = self.communicator?.currentPackage != nil
-        debugPrint("The transfer is busy: \(transferIsBusy)")
+        let transferIsBusy = self.communicator.currentPackage != nil
+        print("The transfer is busy: \(transferIsBusy)")
         return transferIsBusy
     }
     
@@ -293,7 +332,7 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
         self.sidId = sidId
         self.blobData = blobData
         self.blobCounter = blobCounter
-        self.communicator?.connectToSid(sidId)
+        self.communicator.connectToSid(sidId)
     }
     
     /**
@@ -316,16 +355,16 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
      Sending heartbeats message to SID
      */
     func startSendingHeartbeat() {
-        //debugPrint("sending heartbeat!")
         let message = SIDMessage(id: SIDMessageID.heartBeatRequest, payload: MTUSize())
-        self.sendMessage(message)
+        // TODO: handle error
+        let _ = self.sendMessage(message)
     }
     
     /**
      check out connection state if timer for checkheartbeat response fired
      */
     func checkoutHeartbeatsResponse() {
-        debugPrint("check heartbeats Response!")
+        print("check heartbeats Response!")
         if (self.lastHeartbeatResponseDate.timeIntervalSinceNow + self.heartbeatTimeout/1000) < 0 {
             self.currentConnectionState = .notConnected
         }
@@ -336,9 +375,9 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
      Disconnects from current sid
      */
     open func disconnect() {
-        debugPrint("COM-Manager will be disconnected!")
-        self.communicator?.resetCurrentPackage()
-        self.communicator?.resetFoundSids()
+        print("COM-Manager will be disconnected!")
+        self.communicator.resetCurrentPackage()
+        self.communicator.resetFoundSids()
         self.currentEncryptionState = .shouldEncrypt
         self.cryptoManager = ZeroSecurityManager()
         self.sendHeartbeatsTimer?.invalidate()
@@ -362,7 +401,7 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
      - returns: When already in list it returns true, otherwise false.
      */
     open func hasSidID(_ sidId: String) -> Bool {
-        return (self.communicator?.hasSidID(sidId))!
+        return self.communicator.hasSidID(sidId)
     }
     
     /**
@@ -370,19 +409,19 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
      */
     fileprivate func establishCrypto() {
         if self.sidId.isEmpty || self.sidAccessKey.isEmpty {
-            debugPrint("Not found sidId or access key for cram")
+            print("Not found sidId or access key for cram")
             return
         }
-        self.challenger = BLEChallengeService(deviceId: self.deviceId, sidId: self.sidId, leaseTokenId: self.leaseTokenId, sidAccessKey: self.sidAccessKey)
+        self.challenger = BLEChallengeService(leaseId: self.leaseId, sidId: self.sidId, leaseTokenId: self.leaseTokenId, sidAccessKey: self.sidAccessKey)
         self.challenger?.delegate = self
         if self.challenger == nil {
-            debugPrint("Cram could not be initialized")
+            print("Cram could not be initialized")
             return
         }
         do {
             try self.challenger?.beginChallenge()
         } catch {
-            debugPrint("Will be both BLE and Comm. disconnected!")
+            print("Will be both BLE and Comm. disconnected!")
             self.disconnect()
             self.disconnectTransporter()
         }
@@ -416,12 +455,12 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
             }
             
             let message = SIDMessage(id: SIDMessageID.serviceGrant, payload: payload)
-            self.sendMessage(message)
+            let _ = self.sendMessage(message)
             
             if let stop = stopPayload {
                 Delay(0.5, closure: { () -> () in
                     let message = SIDMessage(id: SIDMessageID.serviceGrant, payload: stop)
-                    self.sendMessage(message)
+                    let _ = self.sendMessage(message)
                 })
             }
         }
@@ -432,7 +471,7 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
      */
     fileprivate func sendMtuRequest() {
         let message = SIDMessage(id: SIDMessageID.mtuRequest, payload: MTUSize())
-        self.sendMessage(message)
+        let _ = self.sendMessage(message)
     }
     
     /**
@@ -443,12 +482,12 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
             if let payload = LTBlobPayload(blobData: blobData!)
             {
                 let message = SIDMessage(id: .ltBlob, payload: payload)
-                self.sendMessage(message)
+                let _ = sendMessage(message)
             } else {
-                debugPrint("Blob data error")
+                print("Blob data error")
             }
         } else {
-            debugPrint("Blob data error")
+            print("Blob data error")
         }
     }
     
@@ -462,19 +501,20 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
      - returns:  (success: Bool, error: String?) A Tuple containing a success boolean and a error string or nil
      */
     func sendMessage(_ message: SIDMessage) -> (success: Bool, error: String?) {
-        if self.communicator?.currentPackage != nil {
-            debugPrint("Sending package not empty!! Message \(message.id) will not be sent!!")
-            debugPrint("Package: \(self.communicator?.currentPackage?.message)")
+        if communicator.currentPackage != nil {
+            print("Sending package not empty!! Message \(message.id) will not be sent!!")
+            print("Package: \(communicator.currentPackage?.message)")
             return (false, "Sending in progress")
         } else {
-            let data = self.cryptoManager.encryptMessage(message)
-            self.communicator?.sendData(data)
-            //debugPrint("----------------------------------------")
-//            debugPrint("Send Encrypted Message: \(data.toHexString())")
-//            debugPrint("Same message decrypted: \(self.cryptoManager.decryptData(data).data.toHexString())")
-            //let key = NSData.withBytes(self.cryptoManager.key)
-            //debugPrint("With key: \(key.toHexString())")
-            //debugPrint("-----------  sended message with id: \(message.id) -------------")
+            let data = cryptoManager.encryptMessage(message)
+            return communicator.sendData(data)
+            
+            /*
+            print("Send Encrypted Message: \(data.toHexString())")
+            print("Same message decrypted: \(self.cryptoManager.decryptData(data).data.toHexString())")
+            let key = NSData.withBytes(self.cryptoManager.key)
+            print("With key: \(key.toHexString())")
+            */
         }
         return (true, nil)
     }
@@ -511,7 +551,7 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
         }
         let error = error
         if theStatus == .triggerStatusUnkown {
-            debugPrint("Trigger status unkown!!")
+            print("Trigger status unkown!!")
         }
         self.delegate?.bleDidReceivedServiceTriggerForStatus(theStatus, error: error)
     }
@@ -523,8 +563,7 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
      - parameter message: the message that will be sent to SID peripheral
      */
     func challengerWantsSendMessage(_ message: SIDMessage) {
-        //print ("cram send message!")
-        self.sendMessage(message)
+        let _ = self.sendMessage(message)
     }
     
     /**
@@ -552,8 +591,14 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
     /**
      SID challenger reports to need send Blob to SID peripheral
      */
-    func challengerNeedsSendBlob() {
-       self.sendBlob()
+    func challengerNeedsSendBlob(latestBlobCounter:Int?) {
+        
+        guard latestBlobCounter == nil || blobCounter >= latestBlobCounter!  else {
+            print("Ask user to get latest blob")
+            self.delegate?.blobIsOutdated()
+            return
+        }
+        self.sendBlob()
     }
     
 //MARK: - SIDCommunicatorDelegate
@@ -571,8 +616,7 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
             let pointer = (messageData as NSData).bytes.bindMemory(to: UInt32.self, capacity: messageData.count)
             let count = count
             let buffer = UnsafeBufferPointer<UInt32>(start:pointer, count:count)
-            let array = [UInt32](buffer)
-            //print ("received Data array:\(array) for message id:\(message.id)")
+            _ = [UInt32](buffer)
             
             switch message.id {
             //MTU Size
@@ -592,7 +636,7 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
                 do {
                     try self.challenger?.handleReceivedChallengerMessage(message)
                 } catch {
-                    debugPrint("Will be both BLE and Comm. disconnected!")
+                    print("Will be both BLE and Comm. disconnected!")
                     self.disconnect()
                     self.disconnectTransporter()
                 }
@@ -615,7 +659,7 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
                 }
             }
         }
-        self.communicator?.resetReceivedPackage()
+        self.communicator.resetReceivedPackage()
     }
     
     /**
@@ -628,11 +672,9 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
             self.currentConnectionState = .connected
             self.sendMtuRequest()
         } else {
-            debugPrint("Will be both BLE and Comm. disconnected!")
-            //self.disconnect()
+            print("Will be both BLE and Comm. disconnected!")
             self.currentConnectionState = .notConnected
         }
-        //self.delegate?.bleDidChangedConnectionState(connected)
     }
     
     /**
@@ -651,5 +693,30 @@ open class BLEComManager: NSObject, BLEChallengeServiceDelegate, SIDCommunicator
      */
     func communicatorDidLostSidIds(_ oldSids: [SID]) {
         self.delegate?.bleDidLostSidIds(oldSids)
+    }
+    
+    /**
+     Communicator reports if a connection attempt succeeded
+     
+     - parameter communicator: The communicator object
+     - parameter sid: The SID the connection is made to
+     */
+    func communicatorDidConnectSid(_ communicator: SIDCommunicator, sid: SID) {
+        // TODO: this has to be advanced to cover further communication between device and sid
+        // it is only used for metrics at the moment
+        delegate?.bleDidConnectSid(self, sid: sid)
+    }
+    
+    /**
+     Communicator reports if a connection attempt failed
+     
+     - parameter communicator: The communicator object
+     - parameter sid: The SID the connection should have made to
+     - parameter error: Describes the error
+     */
+    func communicatorDidFailToConnectSid(_ communicator: SIDCommunicator, sid: SID, error: Error?) {
+        // TODO: this has to be advanced to cover further communication between device and sid
+        // it is only used for metrics at the moment
+        delegate?.bleDidFailToConnectSid(self, sid: sid, error: error)
     }
 }