+++ SecureAccessBLE/Classes/Models/Challenger/PhoneToSidChallenge.swift
@@ -21,12 +21,12 @@ struct PhoneToSidChallenge: SIDMessagePayload {
      The Device ID as a string.
      Read-only, can be only set through a initializer.
      */
-    var deviceID: String {
-        let part = self.data.subdata(in: 0..<36)//NSMakeRange(0, 36))
+    var leaseId: String {
+        let part = self.data.subdata(in: 0..<36) // 36 chars
         if let deviceId = NSString(data: part, encoding: String.Encoding.utf8.rawValue) {
             return deviceId as String
         } else {
-            return ""
+            return "" 
         }
     }
     
@@ -35,7 +35,7 @@ struct PhoneToSidChallenge: SIDMessagePayload {
      Read-only, can be only set through a initializer.
      */
     var sidID: String {
-        let part = self.data.subdata(in: 36..<72)//NSMakeRange(36, 36))
+        let part = self.data.subdata(in: 36..<72) // 36 chars
         if let sidId = NSString(data: part, encoding: String.Encoding.utf8.rawValue) {
             return sidId as String
         } else {
@@ -48,7 +48,7 @@ struct PhoneToSidChallenge: SIDMessagePayload {
      Read-only, can be only set through a initializer.
      */
     var leaseTokenID: String {
-        let part = self.data.subdata(in: 72..<152)//NSMakeRange(72, 80))
+        let part = self.data.subdata(in: 72 ..< 108) // 36 chars
         if let sidId = NSString(data: part, encoding: String.Encoding.utf8.rawValue) {
             return sidId as String
         } else {
@@ -59,8 +59,8 @@ struct PhoneToSidChallenge: SIDMessagePayload {
      The challenge that should be send.
      */
     var challenge: [UInt8] {
-        let part = self.data.subdata(in: 152..<178)//NSMakeRange(152, 16))
-        let challenge = part.bytes//.arrayOfBytes()
+        let part = self.data.subdata(in: 108..<124) // 128 bits
+        let challenge = part.bytes
         return challenge
     }
     
@@ -72,10 +72,10 @@ struct PhoneToSidChallenge: SIDMessagePayload {
      - parameter challenge: The challenge as a UInt8 array
      
      */
-    init(deviceID: String, sidID: String, leaseTokenID: String, challenge: [UInt8]) {
+    init(leaseId: String, sidID: String, leaseTokenID: String, challenge: [UInt8]) {
         let data = NSMutableData()
         
-        if let stringData = deviceID.data(using: String.Encoding.utf8) {
+        if let stringData = leaseId.data(using: String.Encoding.utf8) {
             data.append(stringData)
         }
         if let stringData = sidID.data(using: String.Encoding.utf8) {
@@ -88,7 +88,6 @@ struct PhoneToSidChallenge: SIDMessagePayload {
         }
         data.append(challenge, length: challenge.count)
         self.data = data as Data
-        //        debugPrint("data:\(data) with length:\(self.data.length)")
     }
     
 }